Refactor `transformers` module structure (#120)
* Create __init__.py for transformers module

* Create common utilities for transformers module

* Add MultiLabelBinarizer transformer wrapper

* Delete ds_utils/transformers.py

Remove old transformers.py file (migrated to package structure)

* Fix ImportError in tests/test_transformers.py and syntax error in source (#121)

* Fix ImportError in tests/test_transformers.py and syntax error in multi_label_binarizer.py

- Updated `tests/test_transformers.py` to import `MultiLabelBinarizerTransformer` from the specific submodule.
- Fixed an unclosed triple-quoted docstring in `ds_utils/transformers/multi_label_binarizer.py`.
- Applied linting and formatting.


* Fix ImportError in tests/test_transformers.py and syntax error in multi_label_binarizer.py

- Updated `tests/test_transformers.py` to import `MultiLabelBinarizerTransformer` from the specific submodule.
- Fixed an unclosed triple-quoted docstring in `ds_utils/transformers/multi_label_binarizer.py`.
- Formatted all files with ruff to resolve linting errors.


* Restructure transformer tests and fix ImportError

- Created `tests/test_transformers/` directory with `__init__.py`.
- Moved and renamed `tests/test_transformers.py` to `tests/test_transformers/test_multi_label_binarizer.py`.
- Updated import in the test file to point directly to `ds_utils.transformers.multi_label_binarizer`.
- Fixed a syntax error (unclosed docstring) in `ds_utils/transformers/multi_label_binarizer.py`.
- Applied project-wide formatting and linting.


---------


---------
